requirements.txt
torch>=1.9.0
numpy>=1.19.5
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
gym>=0.18.0
stable-baselines3>=1.0
pyyaml>=5.4.1
